Datascience/src/Database/guides/Rafa_desafio_data/requirements.txt
Flask
Gunicorn
mysql-connector-python==8.0.27
pandas==1.3.4
python-decouple==3.5
psycopg2-binary==2.9.7

